Fixing the bullets
Figured out how to delete the bullets after they collide. I was using the wrong, method, I was using OnCollisionEnter2D, but I should have been using OnCollisionStay2D.  Also turnsout I had an error in sending the bullets towards the player. It was shooting only towards 0,0 because the player object the bullets were referencing the Player prefab's position. To fix this, I removed the Player prefab totally, and moved all my code in the bullet's start to in the shoot method of the shootercontroller. It is working for when the players position has a positive y, but it isn't working with a negative y. I dont know why yet.