Fixed Most things about shooter
I fixed the error in the shooter which was causing the bullets to be missing if the player was not in the quadrant to the upper left of the enemy. This was caused by a copy and paste mistake, rather than comparing player and bullet's position, I was comparing bullet and shooter's position. Added code so that the shooter retreats after the player gets close to it.  Also added locations for shooters to shoot bullets from, from their upper lower left and right hand corners