Updating enemies to attack player
After randomly generating the board, none of the objects that need a player object are referencing it, so they dont move according the the player. I tried to fix the camera movement for the first part of the period, but got stuck. I switched to activating the player variable on the enemies so that they now attack the player like they used to. This worked, but I had to make the player a static variable.