Broke and Fixed stuff
Now instead of instantiating fomr the camera, I am instantiating from a loader object. I ended up breaking a bunch of things inthe process, so I spent the period fixing them. Now I have the camera attaching to the player, but not to the player's center, rather where the player is in comparison to tje camera